# openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/health":{"get":{"summary":"Health Check","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object","title":"Response Health Check Health Get"}}}}}}},"/test":{"get":{"summary":"Browser Test Page","operationId":"browser_test_page_test_get","responses":{"200":{"description":"Successful Response","content":{"text/html":{"schema":{"type":"string"}}}}}}},"/diagrams/generate":{"post":{"summary":"Generate Diagram","operationId":"generate_diagram_diagrams_generate_post","parameters":[{"name":"x-api-key","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-Api-Key"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DiagramGenerateRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DiagramResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/diagrams/{diagram_id}":{"get":{"summary":"Get Diagram Route","operationId":"get_diagram_route_diagrams__diagram_id__get","parameters":[{"name":"diagram_id","in":"path","required":true,"schema":{"type":"string","title":"Diagram Id"}},{"name":"x-api-key","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-Api-Key"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DiagramResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/preview/{diagram_id}":{"get":{"summary":"Preview Diagram","operationId":"preview_diagram_preview__diagram_id__get","parameters":[{"name":"diagram_id","in":"path","required":true,"schema":{"type":"string","title":"Diagram Id"}}],"responses":{"200":{"description":"Successful Response","content":{"text/html":{"schema":{"type":"string"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/download/{diagram_id}":{"get":{"summary":"Download Diagram","operationId":"download_diagram_download__diagram_id__get","parameters":[{"name":"diagram_id","in":"path","required":true,"schema":{"type":"string","title":"Diagram Id"}}],"responses":{"200":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/diagrams/{diagram_id}/export/all":{"get":{"summary":"Export Diagram All Formats","operationId":"export_diagram_all_formats_diagrams__diagram_id__export_all_get","parameters":[{"name":"diagram_id","in":"path","required":true,"schema":{"type":"string","title":"Diagram Id"}},{"name":"x-api-key","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-Api-Key"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"DiagramGenerateRequest":{"properties":{"description":{"type":"string","title":"Description"},"diagram_type":{"type":"string","title":"Diagram Type","default":"mermaid"},"format":{"type":"string","title":"Format","default":"svg"}},"type":"object","required":["description"],"title":"DiagramGenerateRequest"},"DiagramResponse":{"properties":{"diagram_id":{"type":"string","title":"Diagram Id"},"prompt":{"type":"string","title":"Prompt"},"diagram_type":{"type":"string","title":"Diagram Type"},"format":{"type":"string","title":"Format"},"rendered_svg":{"type":"string","title":"Rendered Svg"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["diagram_id","prompt","diagram_type","format","rendered_svg","created_at"],"title":"DiagramResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}